requirements.txt
streamlit==1.31.0
flask==3.0.0
flask-cors==4.0.0
onnxruntime-directml
rapidfuzz==3.6.1
audio-recorder-streamlit==0.0.8
matplotlib==3.8.2
requests==2.31.0
ollama==0.1.6
openai-whisper==20231117
soundfile==0.12.1
numpy==1.26.3
groq
pyspellchecker
